Tentative de Nextjs Streaming pour SimilarNodes
CF #757
