+++ app/OsmFeature.tsx
@@ -1,4 +1,4 @@
-import Address, { AddressElement, addressKeys } from '@/components/Address'
+import Address, { addressKeys } from '@/components/Address'
 import ContactAndSocial from '@/components/ContactAndSocial'
 import Emoji from '@/components/Emoji'
 import OsmLinks from '@/components/OsmLinks'
@@ -11,23 +11,20 @@ import Tags, {
 import Wikipedia from '@/components/Wikipedia'
 import { omit } from '@/components/utils/utils'
 import languageIcon from '@/public/language.svg'
+import { css } from 'next-yak'
 import Image from 'next/image'
 import GareInfo from './GareInfo'
+import { OsmFeatureHeader, OsmFeatureWrapper } from './OsmFeatureUI'
 import Heritage from './osm/Heritage'
 import { OpeningHours } from './osm/OpeningHours'
 import getName, { getNameKeys, getNames } from './osm/getName'
 import Brand, { Wikidata } from './tags/Brand'
 import Stop, { isNotTransportStop, transportKeys } from './transport/stop/Stop'
 import { computeSncfUicControlDigit } from './utils'
-import { css } from 'next-yak'
-import { OsmFeatureHeader, OsmFeatureWrapper } from './OsmFeatureUI'
+import { Suspense } from 'react'
+import { Loader } from '@/components/loader'
 
-export default function OsmFeature({
-	data,
-	transportStopData,
-	photonAddress,
-	photonFeature,
-}) {
+export default function OsmFeature({ data, transportStopData, photonFeature }) {
 	if (!data.tags) return null
 	const { tags } = data
 
@@ -239,7 +236,15 @@ export default function OsmFeature({
 			<Brand {...{ brand, brandWikidata, brandWikipedia }} />
 			<Tags tags={keyValueTags} />
 			{wikidata && <Wikidata id={wikidata} />}
-			<SimilarNodes node={data} />
+			<Suspense
+				fallback={
+					<Loader>
+						<p>Chargement des lieux similaires</p>
+					</Loader>
+				}
+			>
+				<SimilarNodes node={data} />
+			</Suspense>
 			<OsmLinks data={data} />
 		</OsmFeatureWrapper>
 	)
